* commit '1da9da316ccbfbd9b66159d9f67626dc30a771bb':
  Some fixes
  Some fixes

+++ code/common/Network/test/main.cpp
@@ -157,7 +157,7 @@ TEST(Network, SingleClientUdp)
 
     try
     {
-        std::srand(std::time(0)); // use current time as seed for random generator
+        std::srand(static_cast<unsigned int>(std::time(0))); // use current time as seed for random generator
         unsigned short random_variable = std::rand() % 65000 + 2500;
 
         network::SockAddr clienInfos(random_variable);
@@ -200,7 +200,7 @@ TEST(Network, TwoClientUdp)
 {
     try
     {
-        std::srand(std::time(0)); // use current time as seed for random generator
+        std::srand(static_cast<unsigned int>(std::time(0))); // use current time as seed for random generator
         unsigned short random_variable = std::rand() % 65000 + 2500;
 
         network::SockAddr clienInfos(random_variable);
@@ -263,7 +263,7 @@ TEST(Network, FourClientUdp)
 {
     try
     {
-        std::srand(std::time(0)); // use current time as seed for random generator
+        std::srand(static_cast<unsigned int>(std::time(0))); // use current time as seed for random generator
         unsigned short random_variable = std::rand() % 65000 + 2500;
 
         network::SockAddr clienInfos(random_variable);
@@ -370,7 +370,7 @@ TEST(Network, SingleClientTcp)
 
     try
     {
-        std::srand(std::time(0)); // use current time as seed for random generator
+        std::srand(static_cast<unsigned int>(std::time(0))); // use current time as seed for random generator
         unsigned short random_variable = std::rand() % 65000 + 2500;
         network::SockAddr clienInfos(random_variable);
         network::SockAddr serverInfos(random_variable, "127.0.0.1");
@@ -413,7 +413,7 @@ TEST(Network, TwoClientTcp)
 {
     try
     {
-        std::srand(std::time(0)); // use current time as seed for random generator
+        std::srand(static_cast<unsigned int>(std::time(0))); // use current time as seed for random generator
         unsigned short random_variable = std::rand() % 65000 + 2500;
 
         network::SockAddr clienInfos(random_variable);
@@ -480,7 +480,7 @@ TEST(Network, FourClientTcp)
 {
     try
     {
-        std::srand(std::time(0)); // use current time as seed for random generator
+        std::srand(static_cast<unsigned int>(std::time(0))); // use current time as seed for random generator
         unsigned short random_variable = std::rand() % 65000 + 2500;
 
         network::SockAddr clienInfos(random_variable);
